refactor(restore): single plan body builder shared by CLI and TUI
The restore plan was rendered twice: ShowRestorePlan(logger, config) in
selective.go printed it for the CLI while buildRestorePlanText(config)
built it for the Charm front-end. The two had already drifted -- the CLI
switch had no default arm for an unknown RestoreMode, its TFA/WebAuthn
advisory was a single 140-column line with no separating blank line, and
a nil config made it panic on config.Mode.

Move buildRestorePlanText into its own neutral file so the shared body is
not filed under a TUI entrypoint (which is how the CLI copy drifted in the
first place), and reduce both front-ends to chrome: the CLI keeps its ASCII
banner, the Charm side keeps its styled Pager title.

The CLI now renders the TUI's wrapped two-line TFA advisory and gains the
"Unknown mode" default arm. fmt.Print, not Printf: category paths reach the
body verbatim and a '%' in a path would be consumed as a format verb.
cliWorkflowUI.ShowRestorePlan now returns an error for a nil config instead
of dereferencing it, matching the Charm sibling; a test pins it, as one
already did on the Charm side.

Output stays on os.Stdout. The CLI restore prompts are already split across
stdout and the injected writer, so moving only the plan would not unify
anything; that is deferred to its own change.
